appdelegate: the unread recommend count is four bytes
Diffing the Objective-C ivar metadata of the shipped binary against our
own build turned up seven width mismatches. Six come from NS_ENUM
backings, which the rules exempt. This one does not: the property was a
plain NSInteger while the binary's getter reads *(int *), so its ivar
encoding is i where ours was q.

Retype it to int and make the narrowing explicit where the recommend
callback's wider status flows in, which is the same narrowing the
binary performs when it stores to the four-byte field.

+++ Project/AppDelegate.h
@@ -78,7 +78,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @ghidraAddress 0x5487c (getter)
  * @ghidraAddress 0x5488c (setter)
  */
-@property(nonatomic, assign) NSInteger unreadRecommendCount;
+@property(nonatomic, assign) int unreadRecommendCount;
 /**
  * @brief Whether the Applilink (KONAMI ID) network layer has finished initialising.
  * @ghidraAddress 0x5489c (getter)
